pyproject.toml
[tool.poetry]
name = "prompt-engineer"
version = "0.1.0"
description = ""
authors = ["Marc Mezger <marc.mezger@gmail.com>"]
readme = "README.md"
packages = [{include = "prompt_engineer"}]

[tool.poetry.dependencies]
python = "^3.11"
aleph-alpha-client = "^3.4.1"
langchain = "^0.0.281"
wandb = "^0.15.9"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
loguru = "^0.7.1"
pandas = "^2.1.0"
textstat = "^0.7.3"
spacy = "^3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
